structuring away from 1 big file to smaller files
