+++ src/components/CreateDefaultLearningPath/Table/DroppableItem.tsx
@@ -1,0 +1,25 @@
+import { UniqueIdentifier, useDroppable } from '@dnd-kit/core'
+import { ReactNode } from 'react'
+import { Paper } from '@common/components'
+import { styled } from '@common/theme'
+
+const DroppableContainer = styled(Paper, {
+  shouldForwardProp: (prop) => prop !== 'isover'
+})<{ isover: boolean }>(({ theme, isover }) => ({
+  width: '100%',
+  height: '100%',
+  flexDirection: 'column',
+  border: `2px dashed ${theme.palette.divider}`,
+  backgroundColor: isover ? theme.palette.action.hover : theme.palette.background.paper,
+  borderRadius: theme.shape.borderRadius
+}))
+
+// Right Column Droppable Container
+export const Droppable = ({ id, children }: { id: UniqueIdentifier; children: ReactNode }) => {
+  const { isOver, setNodeRef } = useDroppable({ id })
+  return (
+    <DroppableContainer ref={setNodeRef} isover={isOver} id={id as string}>
+      {children}
+    </DroppableContainer>
+  )
+}
